requirements.txt
altair==5.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
holidays==0.41
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
mdurl==0.1.2
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
plotly==5.18.0
prophet==1.1.5
protobuf==4.25.2
psycopg2-binary==2.9.9
pyarrow==15.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.25
stanio==0.3.0
streamlit==1.30.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0